requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
faker==22.0.0
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.26.3
